Move invocation of 'update' callback.
Rationale for this change:

Before this change, the "update" callback is in a useless place, because
it calls "update" before anything has changed on the Ball.

The question is where to move it to. It doesn't belong in the "then"
branch of the if, because it's not our practice to send update events
after destroy events (see for example lines 200 and 341). So it only
makes sense in the "else" branch.